+++ cli/src/config.rs
@@ -605,6 +605,21 @@ pub fn default_config_migrations() -> Vec<ConfigMigrationRule> {
         ConfigMigrationRule::rename_value("git.auto-local-branch", "git.auto-local-bookmark"),
         // TODO: Delete in jj 0.28+
         ConfigMigrationRule::rename_value("git.push-branch-prefix", "git.push-bookmark-prefix"),
+        // TODO: Delete in jj 0.33+
+        ConfigMigrationRule::rename_update_value(
+            "signing.sign-all",
+            "signing.behavior",
+            |old_value| {
+                if old_value
+                    .as_bool()
+                    .ok_or("signing.sign-all expects a boolean")?
+                {
+                    Ok("own".into())
+                } else {
+                    Ok("keep".into())
+                }
+            },
+        ),
     ]
 }